refactor: pages 폴더 구조로 페이지 컴포넌트 분리
페이지 컴포넌트를 pages 폴더로 통합하여 라우트 단위 컴포넌트와
일반 컴포넌트를 명확히 구분

Changes:
- main/MainPage.tsx → pages/Main/index.tsx
- CreateRoom/CreateRoom.tsx → pages/CreateRoom/index.tsx
- Game/ReadyToGame.tsx → pages/GameRoom/index.tsx
- ReadyToGame 컴포넌트명을 GameRoom으로 변경 (역할 명확화)
- 모든 import 경로 업데이트
