+++ client/src/pages/GameRoom/index.tsx
@@ -1,5 +1,5 @@
-import IconButton from '../components/button/IconButton';
-import FullLayout from '../components/layout/FullLayout';
+import IconButton from '../../components/button/IconButton';
+import FullLayout from '../../components/layout/FullLayout';
 import { useParams, useLocation, useNavigate } from 'react-router-dom';
 import { useRecoilState } from 'recoil';
 import {
@@ -9,37 +9,37 @@ import {
   readyToGameModalState,
   timerCount,
   soundEffectStatus,
-} from '../recoil/recoil';
+} from '../../recoil/recoil';
 import { useEffect, useState, useRef } from 'react';
-import ReadyToGameModal from '../components/modal/ReadyModal';
-import Button from '../components/button/Button';
-import { ChatMessage, GameUserDto, IUserDto, AnswerUserDto } from '../types/dto';
-import ChatRoom from '../components/chatRoom/ChatRoom';
-import CaptainReatyToModal from '../components/modal/CaptainReadyModal';
-import LoadingModal from '../components/modal/LoadingModal';
-import RadioButton from '../components/radioButton/RadioButton';
-import Toast from '../components/toast/Toast';
-import { history } from '../util/history';
-import JoinUser from '../components/ingameComponents/JoinUser';
-import Timer from './timer/Timer';
-import useTimer, { TimerHookProps } from './timer/useTimer';
-import { Realistic } from '../components/party/Realistic';
-import RankingUser from '../components/ingameComponents/RankingUser';
-import DropdownSelect from '../components/dropDown/DropDown';
+import ReadyToGameModal from '../../components/modal/ReadyModal';
+import Button from '../../components/button/Button';
+import { ChatMessage, GameUserDto, IUserDto, AnswerUserDto } from '../../types/dto';
+import ChatRoom from '../../components/chatRoom/ChatRoom';
+import CaptainReatyToModal from '../../components/modal/CaptainReadyModal';
+import LoadingModal from '../../components/modal/LoadingModal';
+import RadioButton from '../../components/radioButton/RadioButton';
+import Toast from '../../components/toast/Toast';
+import { history } from '../../util/history';
+import JoinUser from '../../components/ingameComponents/JoinUser';
+import Timer from '../../Game/timer/Timer';
+import useTimer, { TimerHookProps } from '../../Game/timer/useTimer';
+import { Realistic } from '../../components/party/Realistic';
+import RankingUser from '../../components/ingameComponents/RankingUser';
+import DropdownSelect from '../../components/dropDown/DropDown';
 import { Option } from 'react-dropdown';
-import ReadyStartButton from './RedayStartButton';
-import SliderComponent from '../components/slider/Slider';
-import Wrapper from '../components/Wrapper';
+import ReadyStartButton from '../../Game/RedayStartButton';
+import SliderComponent from '../../components/slider/Slider';
+import Wrapper from '../../components/Wrapper';
 import { motion } from 'framer-motion';
 import 'slick-carousel/slick/slick.css';
 import 'slick-carousel/slick/slick-theme.css';
 import Slider from 'react-slick';
-import { trackEvent } from '../util/googleAnalytics/trackEvent';
-import { GA_EVENT } from '../constants/GA_EVENT';
-import { answerApi, roomApi, userApi } from '../apis';
-import { gameSocket } from '../socket';
+import { trackEvent } from '../../util/googleAnalytics/trackEvent';
+import { GA_EVENT } from '../../constants/GA_EVENT';
+import { answerApi, roomApi, userApi } from '../../apis';
+import { gameSocket } from '../../socket';
 
-const ReadyToGame = () => {
+const GameRoom = () => {
   const params = useParams(); // params를 상수에 할당
   const [, setIsOpen] = useRecoilState(readyToGameModalState);
   const [, setCaptainIsOpen] = useRecoilState(captainReadyToGameModalState);
@@ -1507,4 +1507,4 @@ const ReadyToGame = () => {
   );
 };
 
-export default ReadyToGame;
+export default GameRoom;
