+++ client/src/util/googleAnalytics/PageViewTracker.ts
@@ -1,6 +1,6 @@
-import { useEffect } from "react";
-import { useLocation } from "react-router-dom";
-import ReactGA from "react-ga4";
+import { useEffect } from 'react';
+import { useLocation } from 'react-router-dom';
+import ReactGA from 'react-ga4';
 
 const PageViewTracker = (): null => {
   const location = useLocation();
@@ -9,13 +9,13 @@ const PageViewTracker = (): null => {
     const rawPath = location.pathname;
 
     let page = rawPath;
-    if (rawPath.startsWith("/ReadyToGame/")) page = "/ReadyToGame";
+    if (rawPath.startsWith('/GameRoom/')) page = '/GameRoom';
 
     ReactGA.set({ page });
-    ReactGA.send("pageview");
+    ReactGA.send('pageview');
 
-    if (process.env.NODE_ENV === "development") {
-      console.log("[PAGEVIEW]", page);
+    if (process.env.NODE_ENV === 'development') {
+      console.log('[PAGEVIEW]', page);
     }
   }, [location.pathname]);
